QFunction - compile cuda/hip only functions separately
